requirements.txt
awsiotsdk
psutil
